Do not include userConfig folder in pot sources (#14867)
Fixes #14820

Summary of the issue:
The checkPot script used to check translation comments also checks the subfolder userConfig. This is not correct since the code in userConfig is not NVDA's code but code from add-ons or from the scratchpad.

Description of user facing changes
The checkPot script does not check the userConfig subfolder anymore.

Description of development approach
In addition to COM interface folder, the userConfig folder is also excluded from the check.
Also the exclusion of these two folders is made in the folders' list rather than in the Glob expression since scons' Glob does not support filtering recursively subvolders; this is also probably more efficient to filter the folders than using Glob for each file.