Add-on store: filter legacy add-ons (#14882)
Part of #13985

Summary of the issue:
Legacy add-ons contain invalid metadata, e.g. addonId mismatch.
Invalid metadata can cause problems in NVDA.
For example, if the addonIds don't match, there is no way of tying a add-on store metadata to a running add-on and it's manifest.

Description of user facing changes
Filters legacy add-ons from the add-on store, so that they are no longer accessible.

Description of development approach
Add legacy field to addon details model and filter by it when generating views.