Bugfix when resetting configuration (#14894)
Fixes #13187

Summary of the issue:
After having modified a parameter in a profile and then reset it to its original value with NVDA+control+C, it was not possible to save the configuration with NVDA+control+R; NVDA was complaining that the file system was probably read-only.

Description of user facing changes
NVDA will no longer refuse to save the configuration after a configuration reset.

Description of development approach
Bugfix: when resetting the config, config.conf._dirtyProfiles is cleared

I have also narrowed the exception filtering when saving the config since in #13187, we get a message related to file write error, whereas the real error had nothing to do with it: it was a KeyError when trying to find "notepad" in config.conf._profileCache. I have filtered on PermissionError which is the error that I get when I set nvda.ini (or the whole NVDA setting folder) as read-only and try to save the config.