Fix for the left and right pointing angle brackets, one of the respective typographic quotes (#14940)
Fixes #14939

Summary of the issue:
When the symbol level is set to none, left double angle bracket and right double angle bracket are read, which should not be the case.

Description of user facing changes
The left double angle bracket and right double angle bracket symbols are no longer read, when the symbol level is set to none and speech synthesizers will properly react on these symbols.

Description of development approach
This needed just the change of the english symbols.dic file. It should propagate all locales and languages. It overwrites erroneous definition in dldr.dic for all locales.