Inform the user that the configuration cannot be saved from a temporary version of NVDA (launcher) (#14914)
Summary of the issue:
When running NVDA temp version (from launcher), the configuration cannot be save due to possible compatibility of version scheme. When pressing NVDA+control+C, the configuration is not saved and a message is logged in the log. However, NVDA reports "Configuration saved".

Description of user facing changes
When running the temporary version of the launcher:

NVDA will not report erroneously that the config was saved when trying to save it with NVDA+control+C command.
NVDA will also not offer anymore the possibility to save the configuration in the NVDA menu.
Description of development approach
As when running in secure mode:

When trying to save the config, use the decorator to block action and report that the config cannot be saved in this version of NVDA.
Also removed the "Save configuration" item in the NVDA menu when running the temp version (launcher)