Add-on store: Improve disk management for launcher and secure copies (#14955)
Part of #13985
Originally raised by @XLTechie in #14912 (reply in thread)

Summary of the issue:
The installer for add-on store versions of NVDA, is creating the addonStore folder in %appdata%\nvda, even before the license agreement is accepted!
E.g. If starting the installer, and selecting "exit" instead of agreeing to the license agreement, the addonStore directory is created and populated with some files and directories under %appdata%\nvda.
It won't create %appdata%\nvda, but if that already exists, it will create the store directory.

This is absolutely not expected behavior. NVDA should never be doing any directory and file creation on the local system before "install" is selected.
I have not looked at the code, so do not know what this would do to an existing installed store copy, but really no user would expect that the installed NVDA's configuration could be altered by an installer that was not told to "install".

Description of user facing changes
Ensure the caching behaviour of add-ons matches that of NVDA's config.
Files should only be written to disk when not running as launcher nor in a secure context.

Description of development approach
Adds relevant checks to prevent saving the add-on handler state and add-on store caches to disk.