Improve shortcut key reporting (#14900)
Fixes #10807

Summary of the issue:
When reported by speech, some shortcut key or combination cannot be understood correctly in the following cases:

A punctuation may not be reported due to symbol level being too low, e.g. "alt+." reported as "alt plus" instead of "alt plus dot" in Outlook.
A character reported as a word instead of being reported as a character. E.g. in French, the letter "à" (a grave) is a word; in MS Word, the shortcut "alt+à" is reported as "alt plus a" instead of "alt plus a grave". Same for "Y" which is a word in French and the name of the letter is "i grec" whereas the word just reads "i" ("ee").
Description of user facing changes
Shortcut key reporting will be improved in some cases. This applies to on-demand reporting (shift+numpad2) as well as when "Report Object Shortcut Keys" option is checked in Object presentation settings.

The cases where it should be improved are:

punctuation key, e.g. "." (dot)
letter keys when the name of the letter is not spoken the same way as the equivalent one-letter word, e.g. "à" or "y" in French.
Braille reporting should remain unchanged.

Description of development approach
Generate a speech sequence to report shortcut keys from the shortcut key string provided by the API. This shortcut key string may contain one or two shortcuts. When there are two shortcuts, they are separated by two spaces.
Each shortcut is split into its component keys.
If a key is a punctuation, symbol reporting processing is used so that the punctuation can be reported no matter the punctuation level.
If a key is a single character, CharacterModeCommand is passed to the synth so that the name of the letter is reported and not a possible one-letter word.