Add-on store (#13985)
Summary of the issue:
Historically, NVDA users have had to browse websites or use add-ons, not directly run/developed by NV Access, in order to learn about, download, and install add-ons. For a long time, easier manual updates for add-ons, and automatic updates for add-ons has been desired by NVDA users. Additionally, with the more common and structured approach to add-on API compatibility breaks, lowering the friction for delivery of updated (and compatible) add-ons is more important than ever.

User stories enabled by this work:

As an NVDA user, I want to easily browse add-ons which are compatible with my version of NVDA, to learn if they may fulfill an unmeet user need.
As an NVDA user, I want to be able to install compatible NVDA add-ons that I have determined will fulfill an unmet user need.
Additionally, this work unblocks the following user stories which may yet be incorporated into this PR:

As an NVDA user, I want to be notified when there are updates to add-ons I have installed, so that I can manually update the add-ons when I am ready to do so.
As an NVDA user, I want NVDA to automatically keep add-ons up to date with the latest version available.
When NVDA notifies of an update (to NVDA), I want to be able to check the compatibility status of the add-ons I have installed, so that I can determine whether the NVDA update will break my core requirements.
Description of changes for the user:
A new dialog is introduced (NVDA menu, tools, addon store) which shows the latest versions of add-ons (submitted to the add-on store).

The list of functionality for the add-on store is expansive.
This should be covered in the user guide and manual test plan