Eurobraille bnote and bbook automatic detection (#14690)
Summary of the issue:
The b.note "VID_28AC&PID_0012"/"VID_28AC&PID_0013" is not detected automatically
The b.book "VID_28AC&PID_0020"/"VID_28AC&PID_0021" is not detected automatically
the b.note joystick1Left/joystick1Right should do the action "braille_scrollBack/Forward" instead of "review_next/previousCharacter"
when connected to a b.note/b.book, a message should be send to the braille display with the format {"S", "n", "NVDA/XX"}, where "XX" is the number of the COM-port. A {"S", "n" ""} message should be sent when disconnecting.

Description of user facing changes
NVDA automatically detects the b.note and b.book
some gestures changed when a b.note is connected

Description of development approach
I have added the PD/VID for the b.note and b.book to bdDetect.py for automatic detection.
In the existing eurobraille.py display driver I have added the "deviceType" for the b.note and b.book
In the existing eurobraille.py, when a b.note is connected, another gesture map is used than the default one for the change in joystick1Left/joystick1Right.
I have changed the name to "eurobraille Displays"
In the existing eurobraille.py, when a device is connected and is confirmed to be a b.note or b.book, NVDA sends a identifier for the connection. When NVDA terminates it sends an empty message to reset this identifier to the display.