Add-on store: fix support for starting NVDA with add-ons disabled (#14967)
Raised in #14912 (comment)

Summary of the issue:
When starting NVDA with add-ons disabled (e.g. through --disable-addons), enabled add-ons are not displayed in the add-on store.

Description of user facing changes
When NVDA starts with add-ons disabled:

Add-on store now displays enabled add-ons
A banner is shown to warn the user that add-ons are disabled
Description of development approach
Add a new state to differentiate between enabled and running, and enabled while not running.