Moved the new Add-on Store subchapter to its own chapter in the User Guide (#14971)
Summary of the issue:
The Add-on Store subchapter was long enough to warrant a standalone chapter.

Description of user facing changes
Moved the store subchapter of "Extra Tools", to its own chapter before "Extra Tools".
Consequently, reduced all headings by one numbering level.
Renamed the new chapter from "Add-on Store", to "Add-ons and the Add-on Store".
Added a stub under the tools menu "Add-on store" description in the "Extra Tools" chapter, pointing to the new chapter.
Wrote a new introductory section, placing the description of add-ons first, and description of store second. Mostly based on what was there before, but more focused on the add-ons than the store at first.
Includes proposed rewording of the non-free add-ons sentences.