Add-on Store: Revision of User Guide language for better explanation of free add-ons with paid components (#14982)
Consideration of the language to use in the User Guide, when describing all add-ons as free in the store, while certain add-ons require the purchase of components or licenses in order to do anything.

Description of user facing changes
Revised language.

Description of development approach
Edited the language to be more in line with the comments in #14966. It may still not be as desired, but this provides a context to revise it in.