Fix on the panel's name in the multi-category settings dialog (#14996)
Issue similar to #14607.

Summary of the issue:
In the multi-category settings dialog, the panel takes the name of the corresponding category. This name is reported when tabbing from the category list to the panel. If the category name contains an ampersand, this character is not reported in the panel's name and it acts as an accelerator instead.

This is the case for NVDA Dev & Test Toolbox add-on where the add-on's name is used as category in the multi-category settings dialog and thus also as panel name.

Description of user facing changes
The panel's name will match the category name, even in the case this name contains an ampersand (&) character.

Description of development approach
Double the ampersand in the panel's name.