when searching for text, "not found" message style changed from ERROR to INFORMATION (#14807)
Summary of the issue:
When the Searching for text function does not find the text, NVDA shows a MessageBox.
Now the style of this MessageBox is ERROR. This pull request changes the style of this MessageBox to INFORMATION.
This message is literally not an error. This is information.

This change is proposed with the following considerations in mind:

Messages with the ERROR style have a recognizable appearance and are accompanied by a system error sound. For users, such messages carry negative associations and cause unpleasant emotions. It seems that there is no need to scare users with an ERROR style message in cases of text not found, because this situation is not comparable to situations of real errors in the system.
Other Windows applications do not use ERROR style messages in similar situations. For example, the standard Notepad application shows information about unfound text using messages in the INFORMATION style, and not ERROR. That is, this pull request brings NVDA closer to the unified Windows interface language.
Description of user facing changes
Message is displayed as information not error

Description of development approach
Changed message style, lint, add translator comments