Add space between add-on list and External install button in the add-on store GUI (#15009)
Fixes #15008

Summary of the issue:
In the add-on store GUI, the add-on list and the "Install from external source" button are too close to each other.

Description of user facing changes
There is now more space between the add-ons list and the button.

Description of development approach
Added a spacer between the two controls in the sizer. I have use the hard-coded value 5 as done elsewhere in this dialog.