Encapsulate NVDA state write paths (#15021)
Summary of the issue:
NVDA uses a variety of paths to write to.
These can be hard to discover or track in the code, as write paths are not stored in globally available named symbols.
For example, if a developer wishes to find code instances where NVDA writes to the profile config directory, searching "profile" will yield additional results, e.g. code comments.

Description of user facing changes
N/A

Description of development approach
Created a centralised class to list all available write paths in NVDA.
This improves discoverability of write paths.