Add-on store - Clarify window title and list label (#15054)
Fixes #15042

Summary of the issue:
The title of add-on store window is formatted such as "Add-on Store - All Installed add-ons"
After the dash, the string is a concatenation of the channel and the type of add-ons in the list.

In English,it makes something almost grammatically correct to build a single expression (if we ignore case). But in other languages, this concatenation without any separator may produce stranger results that sound bad.
Also, it has been noted that the type of add-on listed ("Available add-ons" in this example) is much more important than the channel; thus it should come first.
Description of user facing changes
Reworked the window title to create a title such as "Add-on Store - Installed add-ons (All)"
While at it, also removed the channel in the list name as pointed in Modify the add-on store title bar text #15042 (comment), that is, the list's name is now "Available add-ons" instead of "All Available add-ons" so that the list name is less verbose.
Description of development approach
Implemented changes described below.

Note: I have used parenthesis for the channel, but I may have used a second dash as a separator instead, i.e. "Add-on store - Available add-ons - All". Let me know if you prefer this second version.