Fix context help in Add-on Store (#15080)
Calling context help in the add-on store while the focus is in "Other details" or on the tabs does not jump to the add-on store paragraph.
Context help on other GUI controls (add-ons list, action button) is not impacted by this issue since they have dedicated help targets.
Description of user facing changes
Context help works correctly in add-on store, i.e. jumps to the correct paragraph.

Description of development approach
Fix the anchor's name according to what is in the user guide.

Testing strategy:
Tested F1 with a portable version created from the snapshot generated by appVeyor (nvda_snapshot_pr15080-28549,7e72e668.exe) when the focus is in tabs, description field or other details.