Make add-on store accelerator keys label translatable and discoverable (#15103)
Closes #15010
Closes #15011

Summary of the issue:
In add-on store, some accelerator keys are not discoverable nor translatable:

alt+L for the add-on list is hard-coded (thus not modifiable by translators) and is not discoverable, neither visually (no visible label), nor via object navigation on the whole list
alt+O for "Other details" field is not discoverable visually nor with shift+numpad2.
Description of user facing changes
The labels for the list and for "other details" field are made visible and the accelerator key is underlined.
shift+numpad2 allows to report the accelerator key of the field.
Description of development approach
Do not hide anymore the labels of the list and "other details" field
used the same strings for the label of the list and of the tabs; the only difference is that the "&" making a letter an accelerator key is removed when the string is used for the tabs (and for the title).
Used "alt+c" for the "Actions" button to free up "alt+a" for the add-on list label.
Testing strategy:
For "Other details" and for the 4 possible labels of the add-on list:

Visual check
Check that alt+a/alt+o are working, i.e. moving the focus
When tabbing to "Other details" or to empty add-on list, checked that alt+O/alt+A is reported
When in "Other details" or in empty add-on list, shift+numpad2 reports the shortcut key
When moving nav object to the add-on list (not specifically empty), the shortcut key is reported.