Fix Excel cell formatting reporting (#15100)
Fixes #15091

Summary of the issue:
Pr #14984 broke Excel cell formatting reporting in such a way that it would always report all formatting, regardless of whether the formatting changed between cells.

Description of user facing changes
Ensure that cell formatting isn't repeated needlessly.

Description of development approach
When getting the selection in script_changeSelection, override the parent of the selection with self. This ensures that the format field cache on the work sheet will persist when moving through cells.