Fix first SAPI5 utterance after 10 seconds of silence is missed (#15129)
Fixes #15082

Summary of the issue:
The first utterance after around 10 seconds of silence with a SAPI5 synth is not spoken.

Description of user facing changes
Issue does no longer occur.

Description of development approach
Since #14759 we are modifying some window messages in our process. This includes setting the SMTO_ABORTIFHUNG flag. I was able to pinpoint that setting this flag to a message of type WM_APP caused this particular issue with SAPI5. Therefore I excluded this range of messages (WM_APP through 0xBFFF).