Add-on store: fix various upgrading bugs (#15132)
Fixes #14991

Summary of the issue:
There are various bugs with updating add-ons

An externally sourced add-on replacing an add-on store add-on does not correctly flush the add-on store version JSON cache. This causes the outdated add-on version information to be used by the externally installed add-on.
Updating an add-on through via the store does not correctly reflect the state: this allows an add-on to be updated multiple times. Instead the "download, pending install" state must be tracked better.
Updating an add-on externally does not correctly reflect the state: instead "pending removal" is shown, instead of pending install.
Updating an add-on externally is not handled correctly, causing the updated add-on to never be installed correctly
Updating an add-on through the store is not handled correctly, the new bundle should be installed before the previous bundle is marked for removal. This allows the add-on to run install/uninstall tasks in an expected way
Description of user facing changes
Fix various bugs with updating an add-on

Description of development approach
match the addonHandler installation / removal order when updating an add-on through the add-on store
check download status when determining state
Reflect "pending install" rather than "pending removal" for add-ons being updated.
include downloaded, pending installs in the installed add-ons tab
ensure add-ons are detected correctly when removing pending installs