Stop WASAPI streams if they've been inactive for more than a minute. (#15097)
Fixes #14913.

Summary of the issue:
On some systems, when WASAPI is enabled, the system won't automatically go to sleep, even though it goes to sleep fine when WASAPI is disabled.

Description of user facing changes
When WASAPI audio is enabled, the computer now automatically goes to sleep if appropriate.

Description of development approach
When a WASAPI stream isn't stopped, this unfortunately prevents some systems from going to sleep automatically. I think this is on Windows 10 and earlier, but the precise details aren't clear. According to a Microsoft engineer, this is by design.

I originally thought this occurred when a WASAPI stream remained open. However, some testing revealed that stopping the stream seems to be enough to fix this. Previously, we didn't explicitly stop idle streams. Stopping is better than closing because opening a stream after it is closed can introduce delays or truncated audio.

Specific implementation details:

Add a method to stop a WASAPI stream when it is idle.
When feed() is called, track the current time as the last time the stream was definitely active. If there isn't a pending stream idle check, schedule one.
Every 5 seconds, check if there are any playing streams that haven't been active for 10 seconds. If there are any, stop them.
We use a class-wide check to avoid continually resetting timers on every audio chunk. See the docstring for WasapiWavePlayer._idleCheck for details.